footer alignment matching figma
