+++ src/Pages/Alumni/Alumni.css
@@ -59,6 +59,7 @@ body {
 	font-weight: 500;
 	font-size: 18px;
 	margin: 2em;
+	float: right;
 }
 
 .FooterTitle {
