+++ src/Pages/Alumni/Alumni.js
@@ -1,62 +1,9 @@
 import React from "react";
 import About from "./Components/About";
-import { Container, Row, Col, Button } from "react-bootstrap";
+import Footer from "./Components/Footer";
+import Banner from "./Components/Banner";
+import { Row, Col } from "react-bootstrap";
 import alumni_data from "./Alumni.json";
-import header from "../../assets/teamPageBanner.png";
-
-const Footer = () => {
-	if (window.innerWidth < 480)
-		return (
-			<Container style={{ background: "#3E9CF3", padding: "3.5em 3.5em" }}>
-				<Row>
-					<h1 className={"MobileFooterTitle"}>Contact Us!</h1>
-				</Row>
-				<Row>
-					<p className={"MobileFooterBody"}>
-						We are looking for non-profits who are in need of technological
-						solutions, as well as partners interested in supporting us and our
-						cause.
-					</p>
-				</Row>
-				<Row style={{ paddingTop: "1em" }}>
-					<Button
-						style={{ backgroundColor: "#ffffff" }}
-						className="MobileBottomButton"
-					>
-						<a href="mailto:contact@uoftblueprint.org">
-							contact@uoftblueprint.org
-						</a>
-					</Button>
-				</Row>
-				<Container style={{ paddingTop: "3em", textAlign: "center" }}>
-					<div style={{ textAlign: "center" }}>
-						<p className="Copyright">@ Blueprint 2022.</p>
-					</div>
-				</Container>
-			</Container>
-		);
-	return (
-		<div className="BottomContainer">
-			<Row>
-				<Col md="6">
-					<h4 className={"FooterTitle"}>Contact Us!</h4>
-					<p className={"FooterBody"}>
-						We are looking for non-profits who are in need of technological
-						solutions, as well as partners interested in supporting us and our
-						cause.
-					</p>
-				</Col>
-				<Col>
-					<a href="mailto:contact@uoftblueprint.org">
-						<button className="contact-button BottomButton">
-							contact@uoftblueprint.org
-						</button>
-					</a>
-				</Col>
-			</Row>
-		</div>
-	);
-};
 
 const Member = () => {
 	let members = [];
@@ -81,18 +28,7 @@ const Member = () => {
 const Team = () => {
 	return (
 		<div>
-			<div className="BannerContainer">
-				<img src={header} id="banner" />
-				<div className="BannerText">
-					<div>
-						<h4 id="header-title">Meet our Team</h4>
-						<h2 id="header-text">
-							This is us! We’re a group of students passionate about building
-							tech for the greater good.
-						</h2>
-					</div>
-				</div>
-			</div>
+			<Banner />
 			<div className="container">
 				<div className="TeamContainer">
 					<h5 className="HeaderStyle">
